fix: workflows, update README, add Schema Guard module and other changes (#126)
[Add]
- CODEOWNERS (.github/codeowners)
- Schema Guard module - Protects the migration and avoid dead loops on automatic deploys

[Change]
- Documentation Overhaul: `README.md/CLAUDE.md` updated to reflect code and rules
- migration files: updated to detect and work with potential issues or errors
- Workflow: docker.yml - updated workflow to upload image to GitHub registry
- Migration files: all files re-worked to avoid runtime errors on automatic deploys
- tag naming support + night build (workflow)

[Fix]
- bugfix: create free badge would not show when user has the free badge
- bugfix: edit button won't show to some users

+++ database/migrations/2024_07_14_063959_add_event_column_to_activity_log_table.php
@@ -8,15 +8,23 @@ class AddEventColumnToActivityLogTable extends Migration
 {
     public function up()
     {
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->string('event')->nullable()->after('subject_type');
-        });
+        $connection = Schema::connection(config('activitylog.database_connection'));
+
+        if (! $connection->hasColumn(config('activitylog.table_name'), 'event')) {
+            $connection->table(config('activitylog.table_name'), function (Blueprint $table) {
+                $table->string('event')->nullable()->after('subject_type');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->dropColumn('event');
-        });
+        $connection = Schema::connection(config('activitylog.database_connection'));
+
+        if ($connection->hasColumn(config('activitylog.table_name'), 'event')) {
+            $connection->table(config('activitylog.table_name'), function (Blueprint $table) {
+                $table->dropColumn('event');
+            });
+        }
     }
 }
