+++ database/migrations/2025_08_26_141448_create_special_codes_table.php
@@ -4,20 +4,23 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('special_codes', function (Blueprint $table) {
-            $table->id();
-            $table->timestamps();
-            $table->foreignId('event_id')->constrained()->onDelete('cascade');
-            $table->string('code')->unique();
-            $table->string('class_name');
-            $table->json('constructor_data')->nullable();
-        });
+        if (! Schema::hasTable('special_codes')) {
+            Schema::create('special_codes', function (Blueprint $table) {
+                $table->id();
+                $table->timestamps();
+                $table->foreignId('event_id')->constrained()->onDelete('cascade');
+                $table->string('code')->unique();
+                $table->string('class_name');
+                $table->json('constructor_data')->nullable();
+            });
+        }
     }
 
     /**
